pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "deeprank2"
version = "3.0.0"
description = "DeepRank2 is an open-source deep learning framework for data mining of protein-protein interfaces or single-residue missense variants."
readme = "README.md"
requires-python = ">=3.10"
keywords = [
    "graph neural network",
    "convolutional neural network",
    "protein-protein interfaces",
    "missense variants",
    "deep learning",
    "pytorch",
]
authors = [
    { name = "Giulia Crocioni", email = "g.crocioni@esciencecenter.nl" },
    { name = "Coos Baakman", email = "coos.baakman@radboudumc.nl" },
    { name = "Dani Bodor", email = "d.bodor@esciencecenter.nl" },
    { name = "Daniel Rademaker" },
    { name = "Gayatri Ramakrishnan" },
    { name = "Sven van der Burg" },
    { name = "Li Xue" },
    { name = "Daniil Lepikhov" },
]
license = { text = "Apache-2.0 license" }
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "tables >= 3.8.0",
    "numpy >= 1.21.5",
    "scipy >= 1.11.2",
    "h5py >= 3.6.0",
    "networkx >= 2.6.3",
    "matplotlib >= 3.5.1",
    "pdb2sql >= 0.5.1",
    "scikit-learn >= 1.0.2",
    "chart-studio >= 1.1.0",
    "biopython >= 1.81",
    "python-louvain >= 0.16",
    "markov-clustering >= 0.0.6.dev0",
    "tqdm >= 4.63.0",
    "freesasa >= 2.1.0",
    "tensorboard >= 0.9.0",
    "protobuf >= 3.20.1",
    "ruff >= 0.1.13",
    "dill",
]

[project.optional-dependencies]
# development dependency groups
test = [
    "pytest >= 7.4.0",
    "bump2version",
    "coverage",
    "pycodestyle",
    "pytest-cov",
    "pytest-runner",
    "coveralls",
]
publishing = ["build", "twine", "wheel"]

[project.urls]
Documentation = "https://deeprank2.readthedocs.io/en/latest/?badge=latest"
Repository = "https://github.com/DeepRank/deeprank2"
Changelog = "https://github.com/DeepRank/deeprank2/blob/main/CHANGELOG.md"

[tool.coverage.run]
branch = true
source = ["deeprank2"]

[tool.setuptools.packages.find]
include = ["deeprank2*"]
exclude = ["tests*", "*tests.*", "*tests"]

[tool.setuptools.package-data]
"*" = ["*.xlsx", "*.param", "*.top", "*residue-classes"]

[tool.ruff]
line-length = 159
select = ["ALL"]
ignore = [
    # Unrealistic for this code base
    "PTH",     # flake8-use-pathlib    
    "N",       # naming conventions
    "PLR0912", # Too many branches,
    "PLR0913", # Too many arguments in function definition
    "D102",    # Missing docstring in public method
    # Unwanted
    "FBT",    # Using boolean arguments
    "ANN101", # Missing type annotation for `self` in method
    "ANN102", # Missing type annotation for `cls` in classmethod
    "ANN204", # Missing return type annotation for special (dunder) method
    "B028",   # No explicit `stacklevel` keyword argument found in warning
    "S105",   # Possible hardcoded password
    "S311",   # insecure random generators
    "PT011",  # pytest-raises-too-broad
    "SIM108", # Use ternary operator
    # Unwanted docstrings
    "D100", # Missing module docstring
    "D104", # Missing public package docstring
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in `__init__`
    # Docstring rules irrelevant to the Google style
    "D203", # 1 blank line required before class docstring
    "D204", # 1 blank line required after class docstring
    "D212", # Multi-line docstring summary should start at the first line
    "D213", # Multi-line docstring summary should start at the second line
    "D215", # Section underline is over-indented 
    "D400", # First line should end with a period (clashes with D415:First line should end with a period, question mark, or exclamation point) 
    "D401", # First line of docstring should be in imperative mood
    "D404", # First word of the docstring should not be This
    "D406", # Section name should end with a newline
    "D407", # Missing dashed underline after section 
    "D408", # Section underline should be in the line following the section's name
    "D409", # Section underline should match the length of its name
    "D413", # Missing blank line after last section 
]

# Allow autofix for all enabled rules.
fixable = ["ALL"]
unfixable = ["F401"] # unused imports (should not disappear while editing)

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",    # Use of `assert` detected
    "PLR2004", # Magic value used in comparison
    "D101",    # Missing class docstring
    "D102",    # Missing docstring in public method
    "D103",    # Missing docstring in public function
]
"docs/*" = ["ALL"]
"tests/perf/*" = ["T201"] # Use of print statements

[tool.ruff.lint]
extend-safe-fixes = [
    "D415",   # First line should end with a period, question mark, or exclamation point
    "D300",   # Use triple double quotes `"""`
    "D200",   # One-line docstring should fit on one line
    "TCH",    # Format type checking only imports
    "ISC001", # Implicitly concatenated strings on a single line
    "EM",     # Exception message variables
    "RUF013", # Implicit Optional
    "B006",   # Mutable default argument
]

[tool.ruff.isort]
known-first-party = ["deeprank2"]
